kill jetty.pid only if running

+++ stop.sh
@@ -1,4 +1,7 @@
 #!/bin/bash
 cd ${OPENSHIFT_DATA_DIR}jetty
 
-kill `cat jetty.pid`
+if ps -p `cat jetty.pid` > /dev/null
+then
+   kill `cat jetty.pid`
+fi
